move get_absorber_sites() to super
