+++ larixite/struct/xas.py
@@ -5,13 +5,15 @@
 Atomic structure with an absorber for XAS
 ================================================================
 """
-
+import numpy as np
 from pathlib import Path
 from dataclasses import dataclass
-from typing import Union, Literal
+from typing import Union, Literal, List
 from pymatgen.core import Molecule, Structure, Element, Site
-from larixite.utils import fcompact
+from larixite.utils import fcompact, get_logger
 
+TOIMPLEMENT_ERRMSG = "To implement depending on the structure file format"
+logger = get_logger("larixite.struct")
 
 def site_label(site: Site) -> str:
     """
@@ -32,7 +34,7 @@ def site_label(site: Site) -> str:
 
 @dataclass
 class XasStructure:
-    """Generic data container for an atomic structure model with an absorber"""
+    """Data model for an atomic structure with an absorber"""
 
     name: str  #: unique name, usually the input filename
     label: str  #: a short label, usually the input filename without extension
@@ -45,7 +47,8 @@ class XasStructure:
     radius: float = 7  #: radius of the absorption sphere from the absorbing atom
 
     def __post_init__(self):
-        pass
+        if self.absorber_idx is None:
+            self.absorber_idx = self.get_absorber_sites()[0]
 
     @property
     def cluster_size(self):
@@ -79,6 +82,23 @@ class XasStructure:
             self.absorber_site.coords, self.cluster_size
         )
 
+    @property
+    def sga(self):
+        raise NotImplementedError(TOIMPLEMENT_ERRMSG)
+
+    @property
+    def space_group(self):
+        raise NotImplementedError(TOIMPLEMENT_ERRMSG)
+
+    @property
+    def sym_struct(self):
+        raise NotImplementedError(TOIMPLEMENT_ERRMSG)
+
+    @property
+    def equivalent_sites(self):
+        raise NotImplementedError(TOIMPLEMENT_ERRMSG)
+
+
     def get_site(self, site_index: int):
         return self.struct[site_index]
 
@@ -91,3 +111,34 @@ class XasStructure:
         except Exception:
             occupancy = 1
         return occupancy
+
+    def get_idx_in_struct(self, atom_coords):
+        """Get the index corresponding to the given atomic coordinates (cartesian)"""
+        for idx, atom in enumerate(self.struct):
+            if np.allclose(atom.coords, atom_coords, atol=0.001) is True:
+                return idx
+        errmsg = f"atomic coordinates {atom_coords} not found in self.struct"
+        logger.error(errmsg)
+        # raise IndexError(errmsg)
+        return None
+
+    def get_absorber_sites(self) -> List[int]:
+        """Get the indexes of the absorbing atoms in the structure"""
+        absorber_sites = []
+        for i, sites in enumerate(self.equivalent_sites):
+            site = sites[0]
+            if self.absorber.symbol in site.species_string:
+                occupancy = self.get_occupancy(site.species_string)
+                site_index = self.get_idx_in_struct(site.coords)
+                if occupancy != 1:
+                    logger.warning(
+                        f"{self.name}: absorber {self.absorber.symbol} has occupancy of {occupancy} on site {site_index}"
+                    )
+                absorber_sites.append(site_index)
+        if len(absorber_sites) == 0:
+            errmsg = (
+                f"Absorber {self.absorber.symbol} not found in structure {self.name}"
+            )
+            logger.error(errmsg)
+            raise AttributeError(errmsg)
+        return absorber_sites
