+++ larixite/struct/xas_xyz.py
@@ -7,7 +7,6 @@ Specialize XasStructure to handle structures from XYZ files
 """
 
 from dataclasses import dataclass
-import numpy as np
 from larixite.struct.xas import XasStructure
 from larixite.utils import get_logger
 
@@ -16,10 +15,6 @@ logger = get_logger("larixite.struct")
 
 @dataclass
 class XasStructureXyz(XasStructure):
-    def __post_init__(self):
-        super().__post_init__()
-        if self.absorber_idx is None:
-            self.absorber_idx = self.get_absorber_sites()[0]
 
     @property
     def sga(self):
@@ -31,32 +26,9 @@ class XasStructureXyz(XasStructure):
 
     @property
     def sym_struct(self):
-        raise ArithmeticError("Symmetrized structure not available for XYZ files")
+        """No symmetrized structure for XYZ, simply return the structure"""
+        return self.struct
 
-    def get_idx_in_struct(self, atom_coords):
-        """Get the index corresponding to the given atomic coordinates (cartesian)"""
-        for idx, atom in enumerate(self.struct):
-            if np.allclose(atom.coords, atom_coords, atol=0.001) is True:
-                return idx
-        errmsg = f"atomic coordinates {atom_coords} not found in self.struct"
-        logger.error(errmsg)
-        # raise IndexError(errmsg)
-        return None
-
-    def get_absorber_sites(self):
-        """Get the indexes of the absorbing atoms in the structure"""
-        absorber_sites = []
-        for i, site in enumerate(self.struct.sites):
-            if self.absorber.name in site.species_string:
-                occupancy = self.get_occupancy(site.species_string)
-                site_index = self.get_idx_in_struct(site.coords)
-                if occupancy != 1:
-                    logger.warning(
-                        f"Absorber {self.absorber.name} has occupancy of {occupancy} on site {site_index}"
-                    )
-                absorber_sites.append(site_index)
-        if len(absorber_sites) == 0:
-            errmsg = f"Absorber {self.absorber.name} not found in structure {self.name}"
-            logger.error(errmsg)
-            raise AttributeError(errmsg)
-        return absorber_sites
+    @property
+    def equivalent_sites(self):
+        return [[site] for site in self.struct.sites]
